Struct tiny_xlib::Display

pub struct Display { /* private fields */ }

The display pointer.

Implementations

impl Display

pub fn new(name: Option<&CStr>) -> Result<Self>

Open a new display.

pub unsafe fn from_ptr(ptr: *mut c_void) -> ManuallyDrop<Self>

Create a new Display from a pointer.

Safety

The pointer must be a valid pointer to an Xlib display. In addition, it should only be dropped if the user logically owns the display.

pub fn as_ptr(&self) -> *mut c_void

Get the pointer to the display.